Panel

+++ src/com/girlkun/models/boss/list_boss/doanh_trai/NinjaTim.java
@@ -24,8 +24,8 @@ public class NinjaTim extends Boss {
         super(BossID.NINJA_AO_TIM, new BossData(
                 "Ninja Áo Tím", //name
                 ConstPlayer.TRAI_DAT, //gender
-                new short[]{123, 124, 125, -1, -1, -1}, //outfit {head, body, leg, bag, aura, eff}
-                ((hp / 5)), //dame
+                                new short[]{123, 124, 125, -1, -1, -1}, //outfit {head, body, leg, bag, aura, eff}
+                (int) ((hp / 5)), //dame
                 new double[]{((dame * 350))}, //hp
                 new int[]{57}, //map join
                 new int[][]{
